+++ packages/firebase_ui_localizations/example/android/app/src/main/kotlin/io/flutter/plugins/firebase_ui_localizations_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package io.flutter.plugins.firebase_ui_localizations_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}