+++ packages/firebase_ui_localizations/example/linux/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,15 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <desktop_webview_auth/desktop_webview_auth_plugin.h>
+
+void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) desktop_webview_auth_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "DesktopWebviewAuthPlugin");
+  desktop_webview_auth_plugin_register_with_registrar(desktop_webview_auth_registrar);
+}